+++ src/solvers/dgsem_tree/dg_2d.jl
@@ -523,6 +523,47 @@ function prolong2interfaces!(cache, u, mesh::TreeMesh{2}, equations, dg::DG)
     return nothing
 end
 
+function prolong2interfaces!(cache, u, mesh::TreeMesh{2}, equations,
+                             dg::DG{<:GaussLegendreBasis})
+    @unpack interfaces = cache
+    @unpack orientations, neighbor_ids = interfaces
+    @unpack boundary_interpolation = dg.basis
+    interfaces_u = interfaces.u
+
+    @threaded for interface in eachinterface(dg, cache)
+        left_element = neighbor_ids[1, interface]
+        right_element = neighbor_ids[2, interface]
+
+        if orientations[interface] == 1
+            # interface in x-direction
+            for j in eachnode(dg), v in eachvariable(equations)
+                interfaces_u[1, v, j, interface] = zero(eltype(interfaces_u))
+                interfaces_u[2, v, j, interface] = zero(eltype(interfaces_u))
+                for ii in eachnode(dg)
+                    interfaces_u[1, v, j, interface] += (u[v, ii, j, left_element] *
+                                                         boundary_interpolation[ii, 2])
+                    interfaces_u[2, v, j, interface] += (u[v, ii, j, right_element] *
+                                                         boundary_interpolation[ii, 1])
+                end
+            end
+        else # if orientations[interface] == 2
+            # interface in y-direction
+            for i in eachnode(dg), v in eachvariable(equations)
+                interfaces_u[1, v, i, interface] = zero(eltype(interfaces_u))
+                interfaces_u[2, v, i, interface] = zero(eltype(interfaces_u))
+                for jj in eachnode(dg)
+                    interfaces_u[1, v, i, interface] += (u[v, i, jj, left_element] *
+                                                         boundary_interpolation[jj, 2])
+                    interfaces_u[2, v, i, interface] += (u[v, i, jj, right_element] *
+                                                         boundary_interpolation[jj, 1])
+                end
+            end
+        end
+    end
+
+    return nothing
+end
+
 function calc_interface_flux!(surface_flux_values,
                               mesh::TreeMesh{2},
                               have_nonconservative_terms::False, equations,
@@ -1088,7 +1129,7 @@ function calc_surface_integral!(du, u,
     @unpack surface_flux_values = cache.elements
 
     # This computes the **negative** surface integral contribution,
-    # i.e., M^{-1} * boundary_interpolation^T (which is for DGSEM just M^{-1} * B)
+    # i.e., M^{-1} * boundary_interpolation^T (which is for Gauss-Lobatto DGSEM just M^{-1} * B)
     # and the missing "-" is taken care of by `apply_jacobian!`.
     #
     # We also use explicit assignments instead of `+=` to let `@muladd` turn these
@@ -1116,6 +1157,57 @@ function calc_surface_integral!(du, u,
                 du[v, l, nnodes(dg), element] = (du[v, l, nnodes(dg), element] +
                                                  surface_flux_values[v, l, 4, element] *
                                                  factor)
+            end
+        end
+    end
+
+    return nothing
+end
+
+function calc_surface_integral!(du, u,
+                                mesh::Union{TreeMesh{2}, StructuredMesh{2},
+                                            StructuredMeshView{2}},
+                                equations, surface_integral::SurfaceIntegralWeakForm,
+                                dg::DG{<:GaussLegendreBasis}, cache)
+    @unpack boundary_interpolation_inverse_weights = dg.basis
+    @unpack surface_flux_values = cache.elements
+
+    # This computes the **negative** surface integral contribution,
+    # i.e., M^{-1} * boundary_interpolation^T (which is for Gauss-Legendre DGSEM M^{-1} * L)
+    # and the missing "-" is taken care of by `apply_jacobian!`.
+    #
+    # We also use explicit assignments instead of `+=` to let `@muladd` turn these
+    # into FMAs (see comment at the top of the file).
+    @threaded for element in eachelement(dg, cache)
+        for l in eachnode(dg)
+            for v in eachvariable(equations)
+                for ii in eachnode(dg)
+                    # surface at -x
+                    du[v, ii, l, element] = (du[v, ii, l, element] -
+                                             surface_flux_values[v, l, 1, element] *
+                                             boundary_interpolation_inverse_weights[ii,
+                                                                                    1])
+
+                    # surface at +x
+                    du[v, ii, l, element] = (du[v, ii, l, element] +
+                                             surface_flux_values[v, l, 2, element] *
+                                             boundary_interpolation_inverse_weights[ii,
+                                                                                    2])
+                end
+
+                for jj in eachnode(dg)
+                    # surface at -y
+                    du[v, l, jj, element] = (du[v, l, jj, element] -
+                                             surface_flux_values[v, l, 3, element] *
+                                             boundary_interpolation_inverse_weights[jj,
+                                                                                    1])
+
+                    # surface at +y
+                    du[v, l, jj, element] = (du[v, l, jj, element] +
+                                             surface_flux_values[v, l, 4, element] *
+                                             boundary_interpolation_inverse_weights[jj,
+                                                                                    2])
+                end
             end
         end
     end